Share the subscription identity tag between both generators
The tag taxonomy (which module kind tags with which identity, and the name
the generated code binds it to) is one language-independent fact, so it moves
to generator/types.rs next to PairTopicConsumerKind; each generator renders
only the type name in its own syntax.

With every inner subscription now yielding a (tag, message) pair, the Rust
generator's per-kind recv_tokens carried no information beyond the binding
name the tag already supplies, so build_subscription_struct emits the receive
step itself and the second match on the subscription kind is gone.

Name the discarded peer tag for what it holds in the two hub fixtures that
read a pairing subscription, and route the Python pairing snapshots through
assert_contains_all like their Rust twins.
