+++ crates/generator-internal/src/generator/python/tests/pairings.rs
@@ -80,22 +80,22 @@ fn peer_consumed_topic_wraps_subscribe_peer_without_binding_slots() {
     assert_eq!(artifact.kind, InterfaceKind::PeerConsumedTopic);
 
     let code = &artifact.code_output;
-    for needle in [
-        "node_runner.subscribe_peer(",
-        "LINK_ID,",
-        "PAIRING_NAME,",
-        "PAIRING_TAG,",
-        "TOPIC_NAME,",
-        "class Subscription:",
-        // The tagged pair: the same PeerInfo that paired() returns.
-        "async def next(self) -> Optional[Tuple[peppylib.PeerInfo, Message]]:",
-        "peer, raw_message = item",
-        "return peer, message",
-        "async def wait_paired(",
-        "_deserialize_payload",
-    ] {
-        assert!(code.contains(needle), "missing `{needle}` in:\n{code}");
-    }
+    assert_contains_all(
+        code,
+        &[
+            "node_runner.subscribe_peer(",
+            "LINK_ID,",
+            "PAIRING_NAME,",
+            "PAIRING_TAG,",
+            "TOPIC_NAME,",
+            "class Subscription:",
+            // The tagged pair: the same PeerInfo that paired() returns.
+            "async def next(self) -> Optional[Tuple[peppylib.PeerInfo, Message]]:",
+            "return peer, message",
+            "async def wait_paired(",
+            "_deserialize_payload",
+        ],
+    );
     assert!(
         !code.contains("bound_producer") && !code.contains("TopicMessenger.subscribe("),
         "peer subscriptions must ride subscribe_peer, not the binding-slot path:\n{code}"
@@ -119,21 +119,21 @@ fn observed_consumed_topic_yields_observed_source_tagged_pairs() {
     assert_eq!(artifact.kind, InterfaceKind::ObservedTopic);
 
     let code = &artifact.code_output;
-    for needle in [
-        "node_runner.subscribe_observed(",
-        "class Subscription:",
-        // The tagged pair: the full member identity, so members sharing one
-        // instance stay distinct.
-        "async def next(self) -> Optional[Tuple[peppylib.ObservedSource, Message]]:",
-        "source, raw_message = item",
-        "return source, message",
-        "async def __anext__(self) -> Tuple[peppylib.ObservedSource, Message]:",
-        // The multi-cardinality slot accessor speaks the same type.
-        "def sources(",
-        "_deserialize_payload",
-    ] {
-        assert!(code.contains(needle), "missing `{needle}` in:\n{code}");
-    }
+    assert_contains_all(
+        code,
+        &[
+            "node_runner.subscribe_observed(",
+            "class Subscription:",
+            // The tagged pair: the full member identity, so members sharing one
+            // instance stay distinct.
+            "async def next(self) -> Optional[Tuple[peppylib.ObservedSource, Message]]:",
+            "return source, message",
+            "async def __anext__(self) -> Tuple[peppylib.ObservedSource, Message]:",
+            // The multi-cardinality slot accessor speaks the same type.
+            "def sources(",
+            "_deserialize_payload",
+        ],
+    );
     assert!(
         !code.contains("ProducerRef"),
         "an observer module tags every message with ObservedSource:\n{code}"
